refactor(api): rename rest::v1 -> rest::canonical (live /api/v2 surface, misnamed)
The api-crate module `proximadb_api::rest::v1` holds the LIVE canonical /api/v2 REST surface — port-backed router builders (graph/document/observability/multimodal/hybrid) + handlers wired into production at src/network/rest/v1/handlers.rs:1460-1542. The `v1` path segment is a historical misnomer (the dir uses v1 proto request types internally but serves /api/v2), and it directly caused TD-V1SUNSET-1 + a memory entry to wrongly classify this live code as 'deprecated -> delete' (near-miss caught and reconciled in #825).

Rename the module to rest::canonical so the path declares its true role. Behavior-preserving: the public API surface (proximadb_api::rest::{...}, proximadb_api::{...}) is UNCHANGED — external code reaches these types via short re-export names, not the v1 path (measured: 0 external code imports of the path; only 2 doc-comments, updated). Touch-points: git mv rest/v1 -> rest/canonical (9 files) + rest.rs (pub mod + 3 pub-use sources) + lib.rs:40 re-export source + 2 doc-comment tweaks in src/network/rest/v1/{observability,document}.rs.

Follow-on to TD-V1SUNSET-1 (#825); pre-cleans TD-104 (root-crate decomposition).
