+++ crates/platform/proximadb-api/src/rest/canonical/analytics.rs
