+++ crates/platform/proximadb-api/src/rest/canonical/catalog.rs
