+++ crates/platform/proximadb-api/src/rest/canonical/document.rs
