+++ crates/platform/proximadb-api/src/rest/canonical/entities.rs
