+++ crates/platform/proximadb-api/src/rest/canonical/hybrid.rs
