+++ crates/platform/proximadb-api/src/rest/canonical/mod.rs
