+++ crates/platform/proximadb-api/src/rest/canonical/multimodal_query.rs
