+++ crates/platform/proximadb-api/src/rest/canonical/observability.rs
